+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- [#3505](https://github.com/ignite/cli/pull/3505) Auto migrate dependency tools
+
 ### Changes
 
 - [#3444](https://github.com/ignite/cli/pull/3444) Add support for ICS chains in ts-client generation
@@ -9,6 +13,7 @@
 
 ### Fixes
 
+- [#3497](https://github.com/ignite/cli/pull/3497) Use corret bank balance query url in faucet openapi
 - [#3481](https://github.com/ignite/cli/pull/3481) Use correct checksum format in release checksum file
 - [#3470](https://github.com/ignite/cli/pull/3470) Prevent overriding minimum-gas-prices with default value