+++ changelog.md
@@ -6,6 +6,7 @@
 
 - [#3444](https://github.com/ignite/cli/pull/3444) Add support for ICS chains in ts-client generation
 - [#3494](https://github.com/ignite/cli/pull/3494) bump `cosmos-sdk` and `cometbft` versions
+- [#3529](https://github.com/ignite/cli/pull/3529) Refactor plugin system to use gRPC
 
 ### Fixes