+++ changelog.md
@@ -5,6 +5,7 @@
 ### Features
 
 - [#3505](https://github.com/ignite/cli/pull/3505) Auto migrate dependency tools
+- [#3536](https://github.com/ignite/cli/pull/3536) Change app.go to v2 and add AppWiring feature 
 
 ### Changes